// src/components/ablum-cover/types.ts
export interface IAblumCoverProps {
  iten: Record<string, any>
  width?: number
  height?: number
  bgp?: number
  mskWidth?: number
  wWidth?: number
}

export interface IAblumCoverWrapper {
  wWidth: number
}

export interface IAblumCoverImg {
  width?: number
  height?: number
  bgp?: number
  mskWidth?: number
}